# Reduce redundant pack reads during building bitmap filter for delta merge case (release-8.5)