--- a/dbms/src/Storages/DeltaMerge/DeltaMerge.h
+++ b/dbms/src/Storages/DeltaMerge/DeltaMerge.h
@@ -150,7 +150,7 @@ public:
         }
         else
         {
-            use_stable_rows = delta_index_it.getSid();
+            use_stable_rows = delta_index_it->getSid();
         }
         auto all_range = RowKeyRange::newAll(is_common_handle, rowkey_column_size);
         last_value = all_range.getStart().toRowKeyValue();
@@ -366,10 +366,10 @@ private:
             }
             else
             {
-                if (delta_index_it.isDelete())
+                if (delta_index_it->isDelete())
                 {
                     // Delete.
-                    writeDeleteFromDelta(delta_index_it.getCount());
+                    writeDeleteFromDelta(delta_index_it->getCount());
                 }
                 else
                 {
@@ -377,17 +377,17 @@ private:
                     bool do_write = true;
                     if constexpr (skippable_place)
                     {
-                        if (delta_index_it.getSid() < sk_skip_stable_rows)
+                        if (delta_index_it->getSid() < sk_skip_stable_rows)
                         {
                             do_write = false;
-                            sk_skip_total_rows += delta_index_it.getCount();
+                            sk_skip_total_rows += delta_index_it->getCount();
                         }
                     }
 
                     if (do_write)
                     {
-                        use_delta_offset = delta_index_it.getValue();
-                        use_delta_rows = delta_index_it.getCount();
+                        use_delta_offset = delta_index_it->getValue();
+                        use_delta_rows = delta_index_it->getCount();
                         writeInsertFromDelta(output_columns, output_write_limit);
                     }
                 }
@@ -627,9 +627,9 @@ private:
     {
         UInt64 prev_sid;
         {
-            prev_sid = delta_index_it.getSid();
-            if (delta_index_it.isDelete())
-                prev_sid += delta_index_it.getCount();
+            prev_sid = delta_index_it->getSid();
+            if (delta_index_it->isDelete())
+                prev_sid += delta_index_it->getCount();
         }
 
         ++delta_index_it;
@@ -641,7 +641,7 @@ private:
         }
         else
         {
-            use_stable_rows = delta_index_it.getSid() - prev_sid;
+            use_stable_rows = delta_index_it->getSid() - prev_sid;
         }
     }
 };