--- a/dbms/src/Storages/DeltaMerge/File/ColumnStream.cpp
+++ b/dbms/src/Storages/DeltaMerge/File/ColumnStream.cpp
@@ -157,7 +157,7 @@ std::unique_ptr<CompressedSeekableReaderBuffer> ColumnReadStream::buildColDataRe
 
     // Try to get the largest buffer size of reading continuous packs
     size_t buffer_size = 0;
-    const auto & pack_res = reader.pack_filter.getPackResConst();
+    const auto & pack_res = reader.pack_filter->getPackResConst();
     for (size_t i = 0; i < n_packs; /*empty*/)
     {
         if (!pack_res[i].isUse())