--- a/dbms/src/Storages/DeltaMerge/File/DMFileBlockInputStream.cpp
+++ b/dbms/src/Storages/DeltaMerge/File/DMFileBlockInputStream.cpp
@@ -58,18 +58,22 @@ DMFileBlockInputStreamPtr DMFileBlockInputStreamBuilder::build(
 
     bool is_common_handle = !rowkey_ranges.empty() && rowkey_ranges[0].is_common_handle;
 
-    DMFilePackFilter pack_filter = DMFilePackFilter::loadFrom(
-        dmfile,
-        index_cache,
-        /*set_cache_if_miss*/ true,
-        rowkey_ranges,
-        rs_filter,
-        read_packs,
-        file_provider,
-        read_limiter,
-        scan_context,
-        tracing_id,
-        read_tag);
+    if (!pack_filter)
+    {
+        auto pack_filter_raw = DMFilePackFilter::loadFrom(
+            dmfile,
+            index_cache,
+            /*set_cache_if_miss*/ true,
+            rowkey_ranges,
+            rs_filter,
+            read_packs,
+            file_provider,
+            read_limiter,
+            scan_context,
+            tracing_id,
+            read_tag);
+        pack_filter = pack_filter_raw.getPackFilterResult();
+    }
 
     bool enable_read_thread = SegmentReaderPoolManager::instance().isSegmentReader();
 
@@ -205,7 +209,7 @@ SkippableBlockInputStreamPtr DMFileBlockInputStreamBuilder::tryBuildWithVectorIn
         enable_del_clean_read,
         is_fast_scan,
         max_data_version,
-        std::move(pack_filter),
+        pack_filter.getPackFilterResult(),
         mark_cache,
         enable_column_cache,
         column_cache,