--- a/dbms/src/Storages/DeltaMerge/File/DMFilePackFilterResult.h
+++ b/dbms/src/Storages/DeltaMerge/File/DMFilePackFilterResult.h
@@ -37,7 +37,8 @@ class DMFilePackFilterResult
 
 public:
     const RSResults & getHandleRes() const { return handle_res; }
-    const RSResults & getPackRes() const { return pack_res; }
+    const RSResults & getPackResConst() const { return pack_res; }
+    RSResults & getPackRes() { return pack_res; }
     UInt64 countUsePack() const;
 
     Handle getMinHandle(