--- a/dbms/src/Storages/DeltaMerge/File/DMFilePackFilter_fwd.h
+++ b/dbms/src/Storages/DeltaMerge/File/DMFilePackFilter_fwd.h
@@ -24,4 +24,7 @@ using IdSetPtr = std::shared_ptr<IdSet>;
 
 class DMFilePackFilter;
 
+class DMFilePackFilterResult;
+using DMFilePackFilterResultPtr = std::shared_ptr<DMFilePackFilterResult>;
+using DMFilePackFilterResults = std::vector<DMFilePackFilterResultPtr>;
 } // namespace DB::DM