Pick #9876 and a part of #9738 to release-8.5
### What problem does this PR solve?

Issue Number: close #9875 

Problem Summary:
According to https://github.com/pingcap/tiflash/pull/9876, this would gain overall **26.5%** Improvement on CHBenchmark 1500 with update workloads.

### What is changed and how it works?

```commit-message
Reduce redundant pack reads during bitmap filter building for delta merge case
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
Improve the performance of `TableScan` by skipping unnecessary row reads
```