@gengliqi 2025-09-26
Before: https://github.com/pingcap/tiflash/commit/dbd76e791de25e3f112efb3f2112e2489cc08872
After: https://github.com/pingcap/tiflash/pull/10293/commits/5758756d4ba6cbd82522f6dc964ce0e4137f1e93
Run CHBenchmark 1500 warehouse.
```sql
mysql> select TIDB_TABLE, sum(DELTA_MEMTABLE_ROWS+DELTA_PERSISTED_ROWS) delta_rows, sum(stable_rows) stable_rows from INFORMATION_SCHEMA.TIFLASH_SEGMENTS where TIDB_DATABASE='chbench_delta' group by TIDB_TABLE order by stable_rows, delta_rows;
+------------+------------+-------------+
| TIDB_TABLE | delta_rows | stable_rows |
+------------+------------+-------------+
| region     |         10 |           0 |
| nation     |         50 |           0 |
| district   |      30000 |           0 |
| supplier   |      36384 |           0 |
| warehouse  |       2712 |         288 |
| item       |          0 |      200000 |
| new_order  |     200772 |    26917020 |
| orders     |    1136056 |    89873752 |
| customer   |    1683216 |    89931328 |
| history    |     917957 |    89987264 |
| stock      |    4821696 |   296328727 |
| order_line |   22772241 |   880956623 |
+------------+------------+-------------+
12 rows in set (0.09 sec)

mysql> select sum(DELTA_MEMTABLE_ROWS+DELTA_PERSISTED_ROWS) delta_rows, sum(stable_rows) stable_rows, 100 * sum(DELTA_MEMTABLE_ROWS+DELTA_PERSISTED_ROWS) / sum(stable_rows) delta_percent from INFORMATION_SCHEMA.TIFLASH_SEGMENTS where
TIDB_DATABASE='chbench_delta';
+------------+-------------+---------------+
| delta_rows | stable_rows | delta_percent |
+------------+-------------+---------------+
|   31601094 |  1474195002 |        2.1436 |
+------------+-------------+---------------+
1 row in set (0.08 sec)
```

## Overall 27% Improvement
Basically the same as the test results of https://github.com/pingcap/tiflash/pull/9876.

Query | Before | After | Improvement
-- | -- | -- | --
Q1 | 3.15 | 1.27 | 248%
Q2 | 1.75 | 0.95 | 184%
Q3 | 4.39 | 1.9 | 231%
Q4 | 5.14 | 2.96 | 174%
Q5 | 14.65 | 11.99 | 122%
Q6 | 3.01 | 1 | 301%
Q7 | 8.32 | 5.8 | 143%
Q8 | 4.88 | 2.51 | 194%
Q9 | 43.02 | 40.58 | 106%
Q10 | 33.95 | 31.2 | 109%
Q11 | 1.48 | 0.62 | 239%
Q12 | 5.37 | 3.35 | 160%
Q13 | 0.74 | 0.4 | 185%
Q14 | 4.11 | 2.18 | 189%
Q15 | 14.2 | 9.66 | 147%
Q16 | 10.66 | 10 | 107%
Q17 | 6.28 | 2.29 | 274%
Q18 | 31.95 | 30.17 | 106%
Q19 | 0.16 | 0.07 | 229%
Q20 | 4.2 | 1.8 | 233%
Q21 | 10.3 | 5.74 | 179%
Q22 | 0.46 | 0.1 | 460%
**Sum** | **212.17** | **166.54** | **127%**